Rename AccountHolderGatewayImpl -> PostgresAccountHolderGateway
